+++ grid2op/Action.py
@@ -1083,6 +1083,26 @@ class Action(GridObjects):
 
         return self
 
+    def is_ambiguous(self):
+        """
+        Says if the action, as defined is ambiguous or not.
+
+        Returns
+        -------
+        res: ``True`` if the action is ambiguous, ``False`` otherwise.
+
+        info: ``dict`` or not
+            More information about the error. If the action is not ambiguous, it values to ``None``
+        """
+        try:
+            self._check_for_ambiguity()
+            res = False
+            info = None
+        except AmbiguousAction as e:
+            info = e
+            res = True
+        return res, info
+
     def _check_for_ambiguity(self):
         """
         This method checks if an action is ambiguous or not. If the instance is ambiguous, an
@@ -1180,13 +1200,13 @@ class Action(GridObjects):
                 raise UnitCommitorRedispachingNotAvailable("Impossible to use a redispatching action in this "
                                                            "environment. Please set up the proper costs for generator")
 
+            if np.any(self._redispatch[~self.gen_redispatchable] != 0.):
+                raise InvalidRedispatching("Trying to apply a redispatching action on a non redispatchable generator")
+
             if np.any(self._redispatch > self.gen_max_ramp_up):
                 raise InvalidRedispatching("Some redispatching amount are above the maximum ramp up")
             if np.any(-self._redispatch > self.gen_max_ramp_down):
                 raise InvalidRedispatching("Some redispatching amount are bellow the maximum ramp down")
-
-            if np.any(self._redispatch[~self.gen_redispatchable] != 0.):
-                raise InvalidRedispatching("Trying to apply a redispatching action on a non redispatchable generator")
 
             if "prod_p" in self._dict_inj:
                 new_p = self._dict_inj["prod_p"]
@@ -1745,6 +1765,84 @@ class Action(GridObjects):
         return res
 
 
+class TopoAndRedispAction(Action):
+    def __init__(self, gridobj):
+        Action.__init__(self, gridobj)
+        self.authorized_keys = set([k for k in self.authorized_keys if k != "injection"])
+
+        self.attr_list_vect = ["_set_line_status", "_switch_line_status",
+                               "_set_topo_vect", "_change_bus_vect",
+                               "_redispatch"]
+
+    def __call__(self):
+        """
+        Compare to the ancestor :func:`Action.__call__` this type of Action doesn't allow to change the injections.
+        The only difference is in the returned value *dict_injection* that is always an empty dictionary.
+
+        Returns
+        -------
+        dict_injection: ``dict``
+            This dictionary is always empty
+
+        set_line_status: :class:`numpy.ndarray`, dtype:int
+            This array is :attr:`Action._set_line_status`
+
+        switch_line_status: :class:`numpy.ndarray`, dtype:bool
+            This array is :attr:`Action._switch_line_status`
+
+        set_topo_vect: :class:`numpy.ndarray`, dtype:int
+            This array is :attr:`Action._set_topo_vect`
+
+        change_bus_vect: :class:`numpy.ndarray`, dtype:bool
+            This array is :attr:`Action._change_bus_vect`
+
+        redispatch: :class:`numpy.ndarray`, dtype:float
+            Thie array is :attr:`Action._redispatch`
+
+        """
+        if self._dict_inj:
+            raise AmbiguousAction("You asked to modify the injection with an action of class \"TopologyAction\".")
+        self._check_for_ambiguity()
+        return {}, self._set_line_status, self._switch_line_status, self._set_topo_vect, self._change_bus_vect,\
+               self._redispatch
+
+    def update(self, dict_):
+        """
+        As its original implementation, this method allows modifying the way a dictionary can be mapped to a valid
+        :class:`Action`.
+
+        It has only minor modifications compared to the original :func:`Action.update` implementation, most notably, it
+        doesn't update the :attr:`Action._dict_inj`. It raises a warning if attempting to change them.
+
+        Parameters
+        ----------
+        dict_: :class:`dict`
+            See the help of :func:`Action.update` for a detailed explanation. **NB** all the explanations concerning the
+            "injection" part is irrelevant for this subclass.
+
+        Returns
+        -------
+        self: :class:`TopologyAction`
+            Return object itself thus allowing multiple calls to "update" to be chained.
+
+        """
+        self._vectorized = None
+        if dict_ is not None:
+            for kk in dict_.keys():
+                if kk not in self.authorized_keys:
+                    warn = "The key \"{}\" used to update an action will be ignored. Valid keys are {}"
+                    warn = warn.format(kk, self.authorized_keys)
+                    warnings.warn(warn)
+
+            # self._digest_injection(dict_)  # only difference compared to the base class implementation.
+            self._digest_setbus(dict_)
+            self._digest_change_bus(dict_)
+            self._digest_set_status(dict_)
+            self._digest_change_status(dict_)
+            self._digest_redispatching(dict_)
+        return self
+
+
 class TopologyAction(Action):
     """
     This class is model only topological actions.
@@ -1836,8 +1934,6 @@ class TopologyAction(Action):
             self._digest_setbus(dict_)
             self._digest_change_bus(dict_)
             self._digest_set_status(dict_)
-            self._digest_hazards(dict_)
-            self._digest_maintenance(dict_)
             self._digest_change_status(dict_)
         return self